pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my_de_project"
version = "0.2.0"
description = "Production-style batch data engineering project for order ingestion"
requires-python = ">=3.11"
dependencies = [
  "pandas",
  "pyyaml",
  "requests",
  "sqlalchemy",
  "pydantic",
  "click",
  "boto3",
  "kafka-python",
  "prometheus_client",
  "PyMySQL",
  "psycopg2-binary",
  "pyarrow"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff"
]

[project.scripts]
my-project = "my_project.cli:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
